backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
apscheduler==3.10.4
httpx==0.27.0
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
openai==1.50.0
bcrypt==4.2.0
aiosqlite==0.20.0
